Create security audit
This project is to review a scenario involving an IT manager’s scope, goals, and risk assessment report. Then, perform an internal audit by completing a controls and compliance checklist. .I followed a structured approach to assess the security controls and compliance of Botium Toys, an organization specializing in the retail and management of various assets, both digital and physical. The audit process was as follows